make dashboard components & modify things
