+++ resources/js/Pages/Backend/Components/Dashboard/PurchaseView.vue
@@ -1,0 +1,55 @@
+<template> <section class="col-span-2 card card-compact bg-base-100 shadow-xl">
+                <div class="card-body">
+                    <h2 class="card-title">Purchase Overview</h2>
+                    <div class="flex flex-flow flex-wrap gap-2 mt-1">
+                        <div
+                            class="grow flex flex-col justify-center gap-3 p-3 border-e border-gray-200"
+                        >
+                            <img
+                                :src="'/icons/colored/bag.svg'"
+                                class="h-10 w-10 mx-auto"
+                            />
+                            <p class="flex justify-between">
+                                <span>82</span>
+                                <span>Purchase</span>
+                            </p>
+                        </div>
+                        <div
+                            class="grow flex flex-col justify-center gap-3 px-6 py-2 border-e border-gray-200"
+                        >
+                            <img
+                                :src="'/icons/colored/sales.svg'"
+                                class="h-10 w-10 mx-auto"
+                            />
+                            <p class="flex justify-between">
+                                <span>₹ 13,573</span>
+                                <span>Cost</span>
+                            </p>
+                        </div>
+                        <div
+                            class="grow flex flex-col justify-center gap-3 px-6 py-2 border-e border-gray-200"
+                        >
+                            <img
+                                :src="'/icons/colored/sales.svg'"
+                                class="h-10 w-10 mx-auto"
+                            />
+                            <p class="flex justify-between">
+                                <span>5</span>
+                                <span>Cancel</span>
+                            </p>
+                        </div>
+                        <div
+                            class="grow flex flex-col justify-center gap-3 px-6 py-2"
+                        >
+                            <img
+                                :src="'/icons/colored/sales.svg'"
+                                class="h-10 w-10 mx-auto"
+                            />
+                            <p class="flex justify-between">
+                                <span>₹ 17432</span>
+                                <span>Return</span>
+                            </p>
+                        </div>
+                    </div>
+                </div>
+            </section></template>
