+++ resources/js/Pages/Backend/Components/Dashboard/SalesPurchaseChart.vue
@@ -1,0 +1,19 @@
+<template>
+    <section class="card card-compact bg-base-100 shadow-xl">
+        <div class="card-body">
+            <div class="flex justify-between items-center">
+                <h2 class="card-title">Sales & Purchase</h2>
+                <h2 class="flex items-center gap-2 border py-1 px-4 rounded-md">
+                    <span>
+                        <img
+                            :src="'/icons/colored/sales.svg'"
+                            class="w-5 h-5"
+                        />
+                    </span>
+                    <span>Weekly</span>
+                </h2>
+            </div>
+            <div class="mt-1"></div>
+        </div>
+    </section>
+</template>
